Squash of 2: KAN-157 Updated Field to turn red for short password
Added logic that turns the password input red when trying to signup with a password shorter then the min length (currently set to 6 characters).

KAN-157 Improved Signup Form

The Options for Avatar Selection in the signup form were not up to date with the edit avatar form. Made those changes, so now it looks better and includes the right fields at the right sizes.