Squash of 5: KAN-152 Added a New Unit Testing Target
- Created a brand new Unit Testing Target in Xcode to allow all team members to add their tests.

- Added iPhone and iPad as a target for the unit tests to allow running the tests in the simulator.

KAN-152 Added Imports for Tests + Required Protocols

Added very basic imports that are required for unit testing as well as a class and protocol required for testing

KAN-152 Added the Testable Component

Added a class with functions for the testable components.

KAN-152 Added isSignedIn Unit Test

Added a unit test to check if user is signed in.
This is important so we can keep track of it later on, to understand which views should be loaded.

Merged in KAN-152-UNIT-TEST-signed-in-state (pull request #79)

KAN-152 UNIT TEST signed in state

* KAN-152 Added a New Unit Testing Target

- Created a brand new Unit Testing Target in Xcode to allow all team members to add their tests.

- Added iPhone and iPad as a target for the unit tests to allow running the tests in the simulator.

* KAN-152 Added Imports for Tests + Required Protocols

Added very basic imports that are required for unit testing as well as a class and protocol required for testing

* KAN-152 Added the Testable Component

Added a class with functions for the testable components.

* KAN-152 Added isSignedIn Unit Test

Added a unit test to check if user is signed in.
This is important so we can keep track of it later on, to understand which views should be loaded.

* Merge remote-tracking branch 'origin/KAN-152-UNIT-TEST-signed-in-state' into KAN-152-UNIT-TEST-signed-in-state


Approved-by: Kelvin Kelvin Mathew