Squash of 3: Merged in KAN-154-UNIT-TEST-input-length (pull request #81)
KAN-154 UNIT TEST input length

* KAN-154 Added a Test for input string length validation

important test for the signup form in the app, where it handles different errors for different issues such as the "password less than 6 characters", which turns the field red.

* KAN-154 Added a 3rd scenario for the 3rd Unit Test

Added a string that is too long for testing the last edge case.

* KAN-154 Updated 3rd Test

removed the password too long test since it not used, there is no limit for password length anyways... making the code simpler to understand.

Approved-by: Kelvin Kelvin Mathew