Squash of 7: Merged in KAN-185-increase-avatarmodel-readability (PR #82)
KAN-185 increase avatarmodel readability

* KAN-185 Using a Map instead of an Array - Ensures Better scalability

Added a map instead of an array of strings so now when we add more options for the avatar, all that is needed it to edit the size of the map, instead of adding the name of each item into the array.

* KAN-185 Added More BG Options for Avatars

* KAN-185 More Backgrounds

Added more backgrounds including a TXST Old Main Easter Egg One!

- Included those BGs in the AvatarModel too

* KAN-185 Added Avatar Facial Hair Feature

Added 3 facial hair options (for now).

- Added a whole new menu for choosing the facial hairs

- updated layer orders to ensure facial hair isn't overlapping anything i shouldn't.

* KAN-185 Facial Hair is now a field in FireStore

Added logic to ensure Facial Hair selections are saved to the cloud just like all the other fields.
This maintain the user's selection until changed, and ensures consistent selections across app sessions

* KAN-185 Facial Hair Saves on signup form

Fixed a small issue where when you signed up (instead of edited a profile) - the facial hair field was not created on firebase and therefore did not save. It works now.

* KAN-185 Added more Facial Hair + improved layering

Added more facial hair options, as well improved the layering in the options preview to ensure logical ordering.

- Added one new mouth option

Approved-by: Kelvin Kelvin Mathew