Squash of 6: KAN-197-ShareCards-As-video (pull request #102)
KAN-197 ShareCards As video

* KAN-197 Exporting cards as a video works!

Added a file to export the cards as a video that does 2 rotations with the BG, and saves it as a perfect loop.

- Ensured the save as photo is still an option.

* KAN-197 Native ConfirmationDialog Instead of CustomView

Made sure the popup that lets us choose to export the card as image or video uses native Apple dialog. Looks better, and better adapted for iOS users

* KAN-197 Updated Custom Animation for Exporting

Looks like full file commit... but I just renamed the file to better fit what it does. Original name was for spray painting animation, but I decided its too spesific.

- Updated the looks of the export animation window (smaller).

- Updated logic so when sheet is dismissed - export ends.

* KAN-197 POST PR: Lottie Animation

Added a lottie (json) Animation instead of the other animation I had before.
This addresses the changes requested by kelvin on the PR.

* KAN-197 POST PR: Removed Bad Shadows From Video Exported Templates

This is a fix for changes request by Kelvin on the PR.

* KAN-197 POST PR: Updated Card 5 Text based on avg color in the drawing

Yoinked the code from the signedinprofileview and applied it here.

Approved-by: Kelvin Kelvin Mathew